Tune parameters for quicker failovers
Kubernetes node ready (node monitor) timeout is set to 90 secs by default.
Reduce it to 40 secs for our env, to get quicker failover of apps.

In kubevirt we reschedule the pod within 15 secs if remote node is unreachable.
In zedkube we look reduce the loop to 10 sec wait instead of 30 secs to detect the
failover quickly and reschedule the apps.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>