forward disabling the apps logs correctly
The correct path to forward the DisableLogs to disable app logs looks
like this:
DevConfig -> AppInstanceConfig -> DomainConfig -> DomainStatus -> newlog

The link between AppInstanceConfig and DomainConfig in zedmanager was
forgotten, which led to the DisableLogs being always false from
DomainConfig onwards. This resulted in inability to disable the logs
for the apps.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>