DPC verify: keep DPCStateIPDNSWait if we get a bind error
Occasionally, we receive a netlink notification that an IP address has been
assigned to a port (by dhcpcd), but due to some kernel timing issue,
the address isn’t ready for socket use yet. If DPC verification runs immediately,
it may fail with a temporary error like "bind: cannot assign requested address".
Rather than marking the DPC as broken, we keep the DPCStateIPDNSWait and retry
later (until it works or a timeout runs out).
This is rare with IPv4, but quite common with IPv6.

Signed-off-by: Milan Lenco <milan@zededa.com>