Updgrade dhcpcd to version 10.1.0
We need dhcpcd version 10.0.0 at least, especially for IPv6, containing these patches:
  - https://github.com/NetworkConfiguration/dhcpcd/commit/5f6f61cbe3edfc8313d7db63d6ecf1b08d7232f1
    Fixes: https://github.com/NetworkConfiguration/dhcpcd/issues/69
    Without this, NetworkMonitor fails to retrieve DHCPv6/RA info from dhcpcd.
  - https://github.com/NetworkConfiguration/dhcpcd/commit/2b4fe4c12b5d4366ff21fabf3a6c3799f8e4fa53
    Fixes: https://github.com/NetworkConfiguration/dhcpcd/issues/183
    Without this, dhcpcd will not obtain NTP server IPs from a DHCPv6 server.

I tried version 10.0.0, but "dhcpcd --release <interface>" crashes with a segfault.

Version 10.1.0 seems to be stable across all the features that we are using.

Note: After upgrading Alpine to version v3.21 or newer, a sufficiently up-to-date dhcpcd
will be available in the official Alpine repositories, so building it manually will no longer
be necessary.

Signed-off-by: Milan Lenco <milan@zededa.com>