Makefile: Fix eve target
Commit 7b9076f9b824fe94de0b7bae644a01d6b340881a renamed the variable ROOTFS_IMG
to ROOTFS_IMG_BASE, which broke the target eve because of dependencies that
ended to a intermediary symbolic that is created during the build. The creation
of this symlink was failing because of the usage of the old variable ROOTFS_IMG.
This commit fixes this issue and make eve works flawlessly.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>