github/workflows: Build eve (arm64) image on arm64 runners
In the build workflow the linuxkit cache is exported from build stage to
build eve image stage. However, when building in a host different from
the target architecture, some packages (such as cross-compilers) will be
built for both architectures, which can lead to a lot of building time.
In order to avoid this situation this commit switches eve image arm64
builds to arm64 runners as well.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>