make devinfo interval configurable
We are now sending device info to the controller every 10 minutes when
there is no real change. When nothing happens this constitutes the
majority of EVE's outgoing traffic. This change makes the interval
configurable via the global config item timer.deviceinfo.interval, so
that the user can adjust it to their needs.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>