kubevirt: fix BaseOs Update, must check for defer due to drain
The previous combination of states is no longer seen here
which leads to BaseOs update skipping checking for a
needed defer while the drain process blocks to allow
a volume rebuild operation to complete.

As part of this fix the condition was cleaned up
for readability.  In kubevirt, BaseOs updates
which are locally available and being switched to
may be deferred while a drain process completes.

Signed-off-by: Andrew Durbin <andrewd@zededa.com>