Update kubernetes components to  latest stable versions
Move to latest stable release versions of all kubernetes components we care for.

1) k3s to 1.33.3
2) kubevirt v1.6.0
3) longhorn v1.9.1
4) virtctl v1.6.0

This will give enough test cycles for Nov-lts release.

We download the kubevirt and longhorn yaml files and modify for our env.
They are in kube container in /etc/kubevirt-operator.yaml and /etc/lh-cfg-v1.9.1.yaml
There is nothing to review in those.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>