Fix the replicated volume status
A replicated volume will always have a designated node where the volume is created.
So if we are processing doUpdateVol() and the volume is replicated, just set the
volume created state and return to publish the volumestatus.

During failover zedmanager looks for volumestatus to be created before activating the app.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>