.gitignore: Fix pkg/external-boot-image build
The build procedure for the package pkg/external-boot-image was changed
recently and started to use Dockerfile.in to generate the final Dockerfile.
Also, the tarball produced by this package is generated under the same
package directory. This broke the pushing of the image because linuxkit
will always mark the repository as dirty since it considers the Dockerfile
and the tarball as changes in the sources. For instance, in the current
state every time pkg/external-boot-image is built we got a dirty hash:

lfedge/eve-external-boot-image:7e86027ed410a803a1c7d0c9f17dc98447f96c7f-dirty-c5eadef

This can be easily fixed by just adding both Dockerfile and the generated
tarball to .gitignore.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>