Add null key verfication to detect TPM reset attacks
If kernel is compiled CONFIG_TCG_TPM2_HMAC enabled, it provides a way
to detect TPM reset attacks that has happned after kernel boot.

This mechanism relies on creating a primary key in null hierarchy, which
by design has a ephemeral seed so when TPM is reset, the primary key
will change. The kernel expose the key name via sysfs to userspace.

In userspace we create the same key in null hierarchy and compare
the key name with the one provided by kernel. If they don't match,
it means TPM has been reset since kernel boot.

To make sure the TPM is not replying a stale state, we also
certificate the key using a random nonce.

This check is performed before security critical operations like
seal/unseal or PCR quote.

The key certification should be done with collaboration with a remote
verifier, but we do it locally since we don't have that infrastructure
implemented yet.

At the time of writing this, eve kernel doesn't have CONFIG_TCG_TPM2_HMAC
enabled, so this change is not active.

Signed-off-by: Shahriyar Jalayeri <shahriyar@posteo.de>