pubsub ensure subscription persistence consistency
Subscriptions _must match the persistence of their corresponding publications_.
Previously, this consistency was not enforced in the msrv service, potentially
causing issues—especially when using the PubSub memdriver, which requires all
subscriptions to be persistent.

This patch fixes the inconsistency and adds a clarifying comment to help
maintain correct persistence settings in future changes.

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>