get-deps: fix empty platforms
if the list of platforms is empty, linuxkit silently
ignore to build and does not return any build args.

Unfortunately linuxkit does not populate the platforms
if the image has been found in local cache.

The solution is to use an old Yoda-trick and use the force.
If the force is used, then linuxkit sets the platforms
nevertheless.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>