Rename the kubevirt EVE image variant to 'k'
A rename is needed since kubevirt is optional.
The generic short name 'k' was chosen to leave the door
open to have this encompass eve-kvm functionality down
the road once we address the image size issue.

This rename affects:
- Makefile
- GitHub workflows
- docs
- build tags and related scripts
- go conditional build tag
- grub
- some code and comments

The source code files still live in pkg/pillar/kubevirt and
we can rename this later if there is an agreement on it; such a
rename would have no EVE user impact.

Signed-off-by: eriknordmark <erik@zededa.com>