fix eve-k  arm64 builds
It looks like when we compile eve-k (aka kubevirt eve) on arm64 some go modules are bringing in the
position independant code compilation flags and causing linker to build PIE executable.

So zedbox is build as PIE exectuable and that is crashing as soon as it is launched.
I am not sure zedbox is designed to run at random addresses every time. Also arm64 or amd64 kvm eve does not generate
PIE zedbox.

So for eve-k tell linker to use -no-pie flag.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>