Fix condition for detecting P3 partition initialization
Its possible that P3 is not initialized if the EVE installer was
configured to use a zfs persist on 1 or more separate disks.
eg. eve_install_disk=sda eve_persist_disk=sdb.

Condition was originally amended for Eval EVE in this commit:
7b9076f9b824fe94de0b7bae644a01d6b340881a.
Attempt to formalize the two conditions expected here and
document the possible zfs condition.

Signed-off-by: Andrew Durbin <andrewd@zededa.com>