docs: add min & max config params values
In the past we had users setting arbitrary values for config params
that were out of range. This resulted in no changes on EVE side and
confusion on the user side.
Added min & max values in human-readable format to the config params
documentation. This way the user can see the valid range of values
when editing config params.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>