Bump eden version to 1.0.12
This release includes minor updates to the eden framework to remove
reliance on the deprecated ZInfoDevice.Network for network status.
It also adds a new test for the LPS /api/v1/network endpoint.

Additionally, neoeden tests have been temporarily disabled until they
are stabilized and will be reintroduced in a future eden release.

Signed-off-by: Milan Lenco <milan@zededa.com>