[storage-init.sh] recreate /persist correctly for eve-k
If /persist has been wiped out we normally recreate it as ext4
since we do not know what is was before. However, if eve-hv-type
is 'k' we know it should be ZFS. This enables (unsupported) manual
conversion from eve-kvm to eve-k as long as the IMGA/IMGB are 512 MB
or larger.

Signed-off-by: eriknordmark <erik@zededa.com>
(cherry picked from commit fb228907d0852980b5a3d0f41ad045f6bef428ee)