eve-k: reduce idle logging, ignore api outage during failover
The 'looking for cont' log near the top of checkAppsStatus
and checkAppsFailover is run when apps are running outside of
any active management operations or app migration operations.
Avoid logging above Functionf in this state to minimize pillar logs.

Kubevirt operations on the Hypervisor interface which either change
the state of a domain or get state of a domain specific to a host/node
should first call scheduledOnNode which implements vm/container handlers.
Don't call the direct replicaVmiScheduledOnMe or replicaPodScheduledOnMe.

During failover multiple nodes may attempt info on a kubevirt domain.
scheduledOnNode is used to skip operations for nodes where an app is not
running. In cases where scheduledOnMe is unable to determine the scheduling
state of an app: the caller should determine if the error is some type
where the k3s api server may be temporarily unavailable.
If the api server is temporarily unavailable then declare app state UNKNOWN
as BROKEN signals to the controller that no further tries will be attempted.

Remove incorrect error log in getNodeDrainRequester(). Its possible to
meet that error log in normal conditions:
- zedkube startup as single node mode
- zedkube recv edgenodeclusterconfig
- zedkube node delete
- zedkube cluster delete

Signed-off-by: Andrew Durbin <andrewd@zededa.com>
(cherry picked from commit 25631d746a5c30b052b39ecbedb71466bef3871f)