Put device in maintenance mode if no edge node certs
After a device has been reinstalled but reusing the the device
certificate the attempts to publish EdgeNodeCerts will be rejected
by the controller for security reasons since they have changed.
(They are stored in /persist/certs hence are lost and recreated on
a device reinstall.) This makes that unusual condition visible to the
user by putting the device in maintenance mode.

Signed-off-by: eriknordmark <erik@zededa.com>