Put device in maintenance mode if no edge node certs
After a device has been reinstalled but reusing the the device
certificate the attempts to publish EdgeNodeCerts will be rejected
by the controller for security reasons since they have changed.
(They are stored in /persist/certs hence are lost and recreated on
a device reinstall.) This makes that unusual condition visible to the
user by putting the device in maintenance mode.

XXX TBD to update the API with a new MaintenanceModeReason for this case.

Signed-off-by: eriknordmark <erik@zededa.com>
(cherry picked from commit d97a3ed13473b615b182576541fa1af42f51679b)