[16.0] Delete VMPersistentState support for EFI bootloader
When a EFI bootloader is started with persist state, a PVC is created with RWX mode by kubevirt upstream code.
In eve-k we did not support RWX volumes yet since they require nfs server deployed and working.
So since RWX volumes are not in good state, kubevirt virt-launcher pod keeps crashing every few hours which bounces the VMs.
Ideally virt-launcher should error out not segfault if volumes are not accessible, thats a kubevirt bug and I reported this
issue to the community and they are working on it. Until that issue is fixed, disable the persist state.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>
(cherry picked from commit 7667b697fbac8b2ed01ee645af2436814c1e9fd4)