NI Reconciler: Check for duplicate app DisplayName
Normally, deploying two applications with the same DisplayName is not allowed.
However, it is possible for user to replace an existing app definition inside
EdgeDevConfig with a new one that has a different UUID but reuses the same
DisplayName.

Because shutting down the original app takes some time, zedmanager may start
bringing up the new app while the old one is still connected. In this window,
zedrouter may attempt to connect both apps to the same network instance.
Since dnsmasq DNS host files are named after the app DisplayName, both apps would
end up using the same file. Once the obsolete app is disconnected, it removes
the shared host file, breaking name-to-IP resolution for the new app
inside the NI.

To prevent this, zedrouter now disallows multiple apps with the same
DisplayName from being connected to a network instance simultaneously.
In such cases, the new app will be marked with an error until the old app is
fully removed. Once cleanup completes, the retry timer will reconnect the new
app and clear the error state.

Signed-off-by: Milan Lenco <milan@zededa.com>