mkimage-raw-efi: Increase EFI and rootfs partition size for eve-k
EVE just grew over time and now it supports different variants, such as
eve-k and nvidia platform that carry more files (libraries, container
images, etc) to the root file system. However, we still rely on the legacy
limitation of 300MB for amd64 generic platforms in order to allow updates
from old (legacy) versions. A new mechanism to mitigate this limitation is
on the way and it should solve the issue for legacy installations. Moving
forward, we need to create more room for new firmware files, future kernel
with more drivers, libraries, etc, so the partition layout must be changed.

This commit changes the partition layout for eve-k:

1. EFI partition grows from 36MB to 2GB
2. rootfs partition grows from 1GB to 4GB

For now only eve-k is changed, but the same partition layout will be
applied to all variants in the near future.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>