tools/collect-sources.sh: Exclude var/lock from tarball extraction
Commit aa18688626ab8d1fb5298d4107f74de351f76de1 updated runc to fix
critical CVEs. In this new runc image, var/lock is not a file but a
directory instead, which is leading collect-sources.sh fail with the
following error:

tar: var/lock: Cannot unlink: Is a directory

This commit fixes this issue by excluding var/lock (not need for collect
sources) from extraction.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>