Improve robustness of TestSingleDPC
TestSingleDPC could intermittently fail by checking the DPC testing state
before NIM updates it after completing testing, resulting in reading
DPCStateNone instead of the expected DPCStateSuccess.
To avoid this race, switch the check to use Eventually and wait for the
success state, as is already done in the other DPC unit tests.

Signed-off-by: Milan Lenco <milan@zededa.com>