dpcreconciler: require VLAN parent adapter to be configured before VLAN creation
Fix missing dependency between a VLAN subinterface and its parent adapter
being configured for VLAN use.

Previously, the VLAN item only depended on the adapter’s existence and MTU.
In some cases, this allowed reconciliation to attempt VLAN creation before
the parent interface (bridge) was fully created and configured for VLANs,
causing the operation to fail.

This change strengthens the dependency by requiring the adapter to be marked
as a VLAN parent (UsedAsVlanParent) in addition to having a sufficient MTU.
This guarantees correct ordering and prevents premature VLAN subinterface
creation.

Signed-off-by: Milan Lenco <milan@zededa.com>