Copy nvidia device plugin only on nvidia platform
This commit fixes a regression where nvidia device plugin daemon set is copied on
amd64 platforms too. That leads to pod in error state, thought that is not a major issue in itself,
but since the pod is not online, eve-k k3s components installation is stuck waiting for all pods to be ready.
In other words, clean USB install of eve-k will hang forever to get to ready state.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>