.github/workflows: Change docker login and remove pull_request_target
With the recent changes on Git Hub, pull_request_target is always executed
from the default branch, which broke the build for LTS branches. The only
reason that we use pull_request_target on build workflows it's because of
docker login action, that requires access to the repo secrets.

Since we are using our self-hosted runners, we can rely on the fact that
these runners already have a default login account configured on docker.
Thus, we can just issue a "docker login" command to ensure the docker
login.

This commit removes the docker login action and uses just the "docker
login" command to perform the login on the runner. By doing so, we can
remove the requirement of pull_request_target, so the build workflow from
master will not be executed on PRs from other branches.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>